Move terrain rendering into a module
